Don't block on determining topic type
Workers are now immidiately created for all topics passed in as args.
The workers will wait for topics to appear to determine type instead of
blocking the main loop on this call. This means that the node's launch
will not be interrupted if an expected topic does not exist yet.